requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-crispy-forms==1.13.0
django-phonenumber-field==6.0.0
phonenumbers==8.12.39
Pillow==8.4.0
sqlparse==0.4.2
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.9.2
djangorestframework==3.13.1
django-rest-swagger==2.2.0
